MS_User_Identity/requirements.txt
django
djangorestframework
pymysql 
cryptography
djangorestframework-simplejwt
django-cors-headers